Engine namespace refactor
