+++ ArchivialLibraryTests/Engine/StatusEngine.tests.cs
@@ -5,11 +5,12 @@ using ArchivialLibrary.Logging.Mock;
 using System;
 using System.Threading;
 using ArchivialLibrary.ServiceCore;
+using ArchivialLibrary.Engine;
 
-namespace ArchivialLibraryTests.Client
+namespace ArchivialLibraryTests.Engine
 {
     [TestClass]
-    public class ConnectionEngineTests
+    public class StatusEngineTests
     {
         private const string TestConnectionString = "fakedb";
 
@@ -17,55 +18,50 @@ namespace ArchivialLibraryTests.Client
 
         [TestMethod]
         [ExpectedException(typeof(ArgumentNullException))]
-        public void ConnectionEngineConstructorThrowsExceptionWhenNoDatabaseIsProvided()
+        public void StatusEngineConstructorThrowsExceptionWhenNoDatabaseIsProvided()
         {
-            ArchivialLibrary.Client.ConnectionEngine engine =
-                new ArchivialLibrary.Client.ConnectionEngine(null, new MockLogger(), 0, SharedMockedCoreSettings);
+            var engine = new StatusEngine(null, new MockLogger(), 0, SharedMockedCoreSettings);
         }
-
+        
         [TestMethod]
         [ExpectedException(typeof(ArgumentNullException))]
-        public void ConnectionEngineConstructorThrowsExceptionWhenNoLoggerIsProvided()
+        public void StatusEngineConstructorThrowsExceptionWhenNoLoggerIsProvided()
         {
             var db = new SQLServerClientDatabase(TestConnectionString, new MockLogger(), SharedMockedCoreSettings);
 
-            ArchivialLibrary.Client.ConnectionEngine engine =
-                new ArchivialLibrary.Client.ConnectionEngine(db, null, 0, SharedMockedCoreSettings);
+            var engine = new StatusEngine(db, null, 0, SharedMockedCoreSettings);
         }
 
         [TestMethod]
-        public void ConnectionEngineConstructorDoesNotThrowWhenValidArgumentsAreProvided()
+        public void StatusEngineConstructorDoesNotThrowWhenValidArgumentsAreProvided()
         {
             var logger = new MockLogger();
             var db = new SQLServerClientDatabase(TestConnectionString, new MockLogger(), SharedMockedCoreSettings);
 
-            ArchivialLibrary.Client.ConnectionEngine engine =
-                new ArchivialLibrary.Client.ConnectionEngine(db, logger, 0, SharedMockedCoreSettings);
+            var engine = new StatusEngine(db, logger, 0, SharedMockedCoreSettings);
 
             Assert.IsNotNull(engine);
         }
 
         [TestMethod]
-        public void ConnectionEngineCanStartAndStop()
+        public void StatusEngineCanStartAndStop()
         {
             var logger = new MockLogger();
             var db = new SQLServerClientDatabase(TestConnectionString, new MockLogger(), SharedMockedCoreSettings);
 
-            ArchivialLibrary.Client.ConnectionEngine engine =
-                new ArchivialLibrary.Client.ConnectionEngine(db, logger, 0, SharedMockedCoreSettings);
+            var engine = new StatusEngine(db, logger, 0, SharedMockedCoreSettings);
 
             engine.BeginStart();
             engine.BeginStop();
         }
 
         [TestMethod]
-        public void ConnectionEngineTriggersStoppedEventWhenEngineHasStopped()
+        public void StatusEngineTriggersStoppedEventWhenEngineHasStopped()
         {
             var logger = new MockLogger();
             var db = new SQLServerClientDatabase(TestConnectionString, new MockLogger(), SharedMockedCoreSettings);
 
-            ArchivialLibrary.Client.ConnectionEngine engine =
-                new ArchivialLibrary.Client.ConnectionEngine(db, logger, 0, SharedMockedCoreSettings);
+            var engine = new StatusEngine(db, logger, 0, SharedMockedCoreSettings);
 
             var signalStoppedEvent = new AutoResetEvent(false);
